Develop (#34)
* PER-224 initial commit of entry create w attach
Initial create of entry crate with attachment handler.  Updated some files
that were poorly formatted from handler that were used to start the new
handler.

* PER-224 renamed handler and updated changelog

* Update node.xml

Modified a few of the tooltips and label for clarity.

* Update node.xml

Changed display name and description to be more clear.

* PER-145 updated readme and changelog
Converted readme and changelog to md files for better formatting in github
and on community.

* PER-224 Fix the test output file, refactor how attachments are added to the request.

* PER-224 updated ars_entry_create_attachment_ce_v1
Removed '!' from chomp so that the value could be assigned to a variable.

* removed git commit dif from itsm7 inc WI create ce v1 handler

* PER-224 updated ars_entry_create_attachment_ce_v1
updated the variable names for code readability

* PER-224 output final ARS field values when debug logging.

* PER-224 updated content for consistency

* Renamed handlers and updated descriptions.

* Formatting edits to the changelogs

* Rearranged the order of changes in changelog

* Origin/feature/generic (#33)

* PER-239 - New ARS REST API Handler

* Updated changelog.md

* PER-239 minor updates
Updated code format and renamed variables for clarity.  also updated some logging statements.

Co-authored-by: Chad Rehm <chad.rehm@kineticdata.com>

Co-authored-by: Brian Tigges <brian.tigges@kineticdata.com>
Co-authored-by: Jack Boespflug <jack.boespflug@kineticdata.com>